♻️ refactor: 공통 컴포넌트 구조 정리
