+++ package.json
@@ -22,6 +22,7 @@
   "dependencies": {
     "@floating-ui/react-dom": "^2.1.0",
     "@lukemorales/query-key-factory": "^1.3.4",
+    "@radix-ui/react-slot": "^1.1.0",
     "@tanstack/query-core": "^5.37.1",
     "@tanstack/react-query": "^5.37.1",
     "@types/crypto-js": "^4.2.2",
